app/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
flask-SQLalchemy==3.1.1
sqlalchemy==2.0.40
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
Jinja2==3.1.6